Function pallet_beefy::pallet::dispatchables::report_future_block_voting_unsigned

pub fn report_future_block_voting_unsigned<T: Config>(
    equivocation_proof: Box<FutureBlockVotingProof<BlockNumberFor<T>, T::BeefyId>>,
    key_owner_proof: T::KeyOwnerProof,
)

Report future block voting equivocation. This method will verify the equivocation proof and validate the given key ownership proof against the extracted offender. If both are valid, the offence will be reported.

This extrinsic must be called unsigned and it is expected that only block authors will call it (validated in ValidateUnsigned), as such if the block author is defined it will be defined as the equivocation reporter.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::report_future_block_voting_unsigned.