Trait pallet_beefy::pallet::Config

pub trait Config: Config {
    type BeefyId: Member + Parameter + BeefyAuthorityId<Keccak256> + MaybeSerializeDeserialize + MaxEncodedLen;
    type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, Self::BeefyId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>;
    type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount;
    type KeyOwnerIdentification: Parameter;
    type HandleEquivocation: HandleEquivocation<Self>;
    type MaxAuthorities: Get<u32>;
    type OnNewValidatorSet: OnNewValidatorSet<<Self as Config>::BeefyId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type BeefyId: Member + Parameter + BeefyAuthorityId<Keccak256> + MaybeSerializeDeserialize + MaxEncodedLen

Authority identifier type

type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, Self::BeefyId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>

A system for proving ownership of keys, i.e. that a given key was part of a validator set, needed for validating equivocation reports.

type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount

The proof of key ownership, used for validating equivocation reports The proof must include the session index and validator count of the session at which the equivocation occurred.

type KeyOwnerIdentification: Parameter

The identification of a key owner, used when reporting equivocations.

type HandleEquivocation: HandleEquivocation<Self>

The equivocation handling subsystem, defines methods to report an offence (after the equivocation has been validated) and for submitting a transaction to report an equivocation (from an offchain context). NOTE: when enabling equivocation handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type MaxAuthorities: Get<u32>

The maximum number of authorities that can be added.

type OnNewValidatorSet: OnNewValidatorSet<<Self as Config>::BeefyId>

A hook to act on the new BEEFY validator set.

For some applications it might be beneficial to make the BEEFY validator set available externally apart from having it in the storage. For instance you might cache a light weight MMR root over validators and make it available for Light Clients.

type WeightInfo: WeightInfo

Weights for this pallet.