Trait pallet_beefy::pallet::Config

pub trait Config: Config {
    type BeefyId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen;
    type MaxAuthorities: Get<u32>;
    type OnNewValidatorSet: OnNewValidatorSet<<Self as Config>::BeefyId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type BeefyId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen

Authority identifier type

type MaxAuthorities: Get<u32>

The maximum number of authorities that can be added.

type OnNewValidatorSet: OnNewValidatorSet<<Self as Config>::BeefyId>

A hook to act on the new BEEFY validator set.

For some applications it might be beneficial to make the BEEFY validator set available externally apart from having it in the storage. For instance you might cache a light weight MMR root over validators and make it available for Light Clients.